Separate test and tested structures.
